agent-sdk/python/requirements.txt
claude-code-sdk>=0.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
